ajout gamegateway
