+++ apps/pogos-games/src/app.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { ChatModule } from './chat/chat.module';
 import { BlackjackModule } from './blackjack/blackjack.module';
 import { CardsModule } from './cards/cards.module';
 import { ConfigModule } from '@nestjs/config';
+import { ChatModule } from '../../../libs/tools/src/chat/chat.module';
 import { ToolsModule } from '../../../libs/tools/src';
 
 @Module({
