+++ libs/tools/src/game/blackjack.gateway.ts
@@ -1,0 +1,79 @@
+import {
+  OnGatewayConnection,
+  OnGatewayDisconnect,
+  SubscribeMessage,
+  WebSocketGateway,
+  WebSocketServer,
+} from '@nestjs/websockets';
+import { Server, Socket } from 'socket.io';
+import { GameService } from './game.service';
+import { ChatGateway } from '../chat/chat.gateway';
+import { GamePlayerResponse } from './dto/response/game-player-response.interface';
+import { GatewayEventsListener } from './enum/gateway/gateway-events-listener.enum';
+import { GatewayEventEmitter } from './enum/gateway/gateway-event-emitter.enum';
+import { GameCreationRequest } from './dto/request/game-creation-request.class';
+import { GameType } from './enum/game-type.enum';
+
+// process.env.FRONTEND_URL
+export class GameGateway
+  extends ChatGateway
+  implements OnGatewayConnection, OnGatewayDisconnect
+{
+  constructor(protected readonly gameService: GameService) {
+    super();
+  }
+
+  @WebSocketServer()
+  server: Server;
+
+  handleConnection(client: Socket) {
+    console.log('Client connected:', client.id);
+  }
+
+  handleDisconnect(client: Socket) {
+    console.log('Client disconnected:', client.id);
+  }
+
+  protected async sendGameAction(
+    players: string[],
+    gameAction: GamePlayerResponse,
+  ) {
+    players.forEach((playerId) => {
+      this.server
+        .to(playerId)
+        .emit(GatewayEventEmitter.PLAYER_UPDATE, gameAction);
+    });
+  }
+
+  @SubscribeMessage(GatewayEventsListener.END_GAME)
+  async handleEndGame(client: Socket, gameId: string) {
+    const gameClients = await this.gameService.endGame(client, gameId);
+    gameClients.forEach((clientId) => {
+      this.server
+        .to(clientId)
+        .emit(GatewayEventEmitter.GAME_UPDATE, 'game ended');
+      this.server.sockets.sockets.get(clientId).disconnect();
+    });
+  }
+
+  @SubscribeMessage(GatewayEventsListener.CREATE_GAME)
+  async handleCreateGame(client: Socket, request: GameCreationRequest) {
+    const gameId = await this.gameService.createGame(
+      client.id,
+      request.type,
+    );
+    console.log('Game created:', gameId);
+    if (request.type == GameType.SOLO) {
+      const response = await this.gameService.startGame(client.id, gameId);
+      client.emit(GatewayEventEmitter.GAME_UPDATE, response);
+    } else {
+      client.emit(GatewayEventEmitter.GAME_UPDATE, gameId);
+    }
+  }
+
+  @SubscribeMessage(GatewayEventsListener.JOIN_GAME)
+  async handleJoinGame(client: Socket, gameId: string) {
+    await this.gameService.joinGame(gameId, client.id);
+    client.emit(GatewayEventEmitter.GAME_UPDATE, gameId);
+  }
+}
