+++ libs/tools/src/game/dto/request/game-creation-request.class.ts
@@ -1,0 +1,8 @@
+import { GameType } from '../../enum/game-type.enum';
+import { IsNotEmpty } from 'class-validator';
+
+export class GameCreationRequest {
+
+@IsNotEmpty()
+  type: string
+}
