+++ libs/tools/src/game/enum/game-type.enum.ts
@@ -1,0 +1,5 @@
+
+export enum GameType {
+  SOLO = "SOLO",
+  MULTIPLAYER = "MULTIPLAYER"
+}
