+++ libs/tools/src/game/enum/gateway/gateway-event-emitter.enum.ts
